Move tes_the_test scenario to its own file
